Create Scatter plot from CSV in R
In R Programming Language we use plot() function to display scatterplot. It takes eight parameters.

    Syntax: plot(x, y, main, xlab, ylab, xlim, ylim, axes)

    Parameters:

        x: sets variable to be used for horizontal coordinates.
        y: sets variable to be used for vertical coordinates.
        xlab: label for horizontal axis.
        ylab: label for vertical axis.
        main: title of the chart.
        xlim: limits of x for plotting values of x.
        ylim: limits of y for plotting values of y.
        axes: it indicates whether both axes should be drawn on the plot.

To set attributes of scatter plot like color and shape of scatter plot points we use “col” attribute to set the color of scatter plot and to set shape we use “pch”, where pch takes numeric values between 0 and 25.
Creating DataFrame:

Here we are going to create dataframe from this dataset. We use read.csv() to read csv file and store that data in a variable.

Jamal Hussain Shah